back-end/youtubeVideo/generatingAudio.py
import os
from google.cloud import texttospeech
os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = "back-end\\youtubeVideo\\cred.json"

def synthesize_speech(text, output_file="narration.mp3", language_code="en-US", voice_name="en-US-Journey-D"):
    client = texttospeech.TextToSpeechClient()

    input_text = texttospeech.SynthesisInput(text=text)

    voice = texttospeech.VoiceSelectionParams(
        language_code=language_code,
        name=voice_name,
        ssml_gender=texttospeech.SsmlVoiceGender.NEUTRAL
    )

    audio_config = texttospeech.AudioConfig(
        audio_encoding=texttospeech.AudioEncoding.MP3
    )

    response = client.synthesize_speech(input=input_text, voice=voice, audio_config=audio_config)

    with open(output_file, "wb") as out:
        out.write(response.audio_content)
        print(f'Audio content written to file "{output_file}"')

    absolute_path = os.path.abspath(output_file)
    print(f'Output file path: {absolute_path}')

    return absolute_path
